Refactor Operations Files
