+++ packages/bodiless-page/src/Operations/index.ts
@@ -1,0 +1,7 @@
+import createRedirect from './Redirect';
+import verifyPage from './PageVerification';
+
+export {
+  createRedirect,
+  verifyPage,
+};
